Moving hooks to hooks-folder (thanks to @framitdavid)
