+++ src/utils/formComponentUtils.ts
@@ -1,93 +1,19 @@
 import type React from 'react';
 
-import { useAppSelector } from 'src/common/hooks/useAppSelector';
 import { getLanguageFromKey, getParsedLanguageFromText, getTextResourceByKey } from 'src/language/sharedLanguage';
 import printStyles from 'src/styles/print.module.css';
 import { AsciiUnitSeparator } from 'src/utils/attachment';
-import { getOptionLookupKey, getRelevantFormDataForOptionSource, setupSourceOptions } from 'src/utils/options';
 import { getTextFromAppOrDefault } from 'src/utils/textResource';
 import type { ExprResolved } from 'src/features/expressions/types';
-import type { IGridStyling, ISelectionComponent } from 'src/layout/layout';
+import type { IGridStyling } from 'src/layout/layout';
 import type { IPageBreak } from 'src/layout/layout.d';
 import type { IAttachment } from 'src/shared/resources/attachments';
-import type { IComponentValidations, IOption, ITextResource, ITextResourceBindings } from 'src/types';
+import type { IComponentValidations, ITextResource, ITextResourceBindings } from 'src/types';
 import type { ILanguage } from 'src/types/shared';
 import type { AnyItem } from 'src/utils/layout/hierarchy.types';
 
 export interface IComponentFormData {
   [binding: string]: string | undefined;
-}
-
-export function useOptionList(component: ISelectionComponent): IOption[] {
-  const textResources = useAppSelector((state) => state.textResources.resources);
-  const formData = useAppSelector((state) => state.formData.formData);
-  const repeatingGroups = useAppSelector((state) => state.formLayout.uiConfig.repeatingGroups);
-  const options = useAppSelector((state) => state.optionState.options);
-
-  if (component.options) {
-    return component.options;
-  }
-  if (component.optionsId) {
-    const key = getOptionLookupKey({
-      id: component.optionsId,
-      mapping: component.mapping,
-    });
-    return options[key]?.options || [];
-  }
-  if (component.source) {
-    const relevantTextResource = textResources.find((e) => e.id === component.source?.label);
-    const reduxOptions =
-      relevantTextResource &&
-      setupSourceOptions({
-        source: component.source,
-        relevantTextResource,
-        relevantFormData: getRelevantFormDataForOptionSource(formData, component.source),
-        repeatingGroups,
-        dataSources: {
-          dataModel: formData,
-        },
-      });
-    return reduxOptions || [];
-  }
-
-  return [];
-}
-
-/**
- * Utility function meant to convert a value for a selection component to a label/text used in Summary
- *
- * Expected to be called from:
- * @see FormComponent.useDisplayData
- */
-export function useSelectedValueToText(component: ISelectionComponent, value: string) {
-  const textResources = useAppSelector((state) => state.textResources.resources);
-  const optionList = useOptionList(component);
-  const label = optionList.find((option) => option.value === value)?.label;
-
-  if (!label) {
-    return value;
-  }
-
-  return getTextResourceByKey(label, textResources) || value;
-}
-
-/**
- * Utility function meant to convert multiple values for a multi-selection component to an object used in Summary
- *
- * Expected to be called from:
- * @see FormComponent.useDisplayData
- */
-export function useCommaSeparatedOptionsToText(component: ISelectionComponent, value: string) {
-  const textResources = useAppSelector((state) => state.textResources.resources);
-  const optionList = useOptionList(component);
-  const split = value.split(',');
-  const out: { [key: string]: string } = {};
-  split?.forEach((part) => {
-    const textKey = optionList.find((option) => option.value === part)?.label || '';
-    out[part] = getTextResourceByKey(textKey, textResources) || part;
-  });
-
-  return out;
 }
 
 export const getTextResource = (resourceKey: string, textResources: ITextResource[]): React.ReactNode => {
